Implements LazyString
Decryption operations are usually slow. If you have a model with a
couple encrypted fields, loading the record will forcefully trigger
decryption for each one of the fields, even if they are not actually
used.

These changes introduce a LazyString object that is returned by the
providers `decrypt` method. The LazyString is wrapper around the
encrypted String object, and will only trigger decryption if the string
is actually called upon.