Allows setting a fallback key for the PostgresPgp decryptor
This can be used when needing to rotate keys for example. By configuring
a fallback key, you can setup a new key that is used for new
encryptions. For decryptions, it will try to decrypt using the key
first. If that fails, it will also try using the fallback key.