Enforce sqlite3 1.3.0 on activerecord 4.2
ActiveRecord 4.2 requires the use of sqlite 1.3.x for a version.
However, 5.2 allows for 1.3 or 1.4, and ActiveRecord 6.0 specifies 1.4.

As such, the appraisal file for 4.2 needs to ensure that sqlite 1.3.x
will be used. The prior implementation was allowing for anything less
than 2.0, which includes 1.4, so that's what bundler was grabbing.

By specifying more specifically to have the version constraint be `~>
1.3.0`, then this will grab something in the 1.3.x series, which is what
ActiveRecord 4.2 expects.